move files from --tests-- to __tests__
