--- src/--test--/PriceTable.test.js
@@ -1,7 +1,0 @@
-import { render } from './test-utils'; // Import the custom render function
-import PriceTable from '../components/detailspage/PriceTable';
-
-it('renders correctly', () => {
-  const { asFragment } = render(<PriceTable />);
-  expect(asFragment()).toMatchSnapshot();
-});
