--- src/--test--/RadialSeparator.test.js
@@ -1,7 +1,0 @@
-import { render } from './test-utils'; // Import the custom render function
-import RadialSeparator from '../components/detailspage/RadialSeparator';
-
-it('renders correctly', () => {
-  const { asFragment } = render(<RadialSeparator />);
-  expect(asFragment()).toMatchSnapshot();
-});
