+++ src/__tests__/unitTests/AddRoom.test.js
@@ -1,0 +1,7 @@
+import { render } from '../../--test--/test-utils'; // Import the custom render function
+import AddRoom from '../../components/addRoom/AddRoom';
+
+it('renders correctly', () => {
+  const { asFragment } = render(<AddRoom />);
+  expect(asFragment()).toMatchSnapshot();
+});
