+++ src/__tests__/unitTests/RadialSeparator.test.js
@@ -1,0 +1,7 @@
+import { render } from '../../--test--/test-utils'; // Import the custom render function
+import RadialSeparator from '../../components/detailspage/RadialSeparator';
+
+it('renders correctly', () => {
+  const { asFragment } = render(<RadialSeparator />);
+  expect(asFragment()).toMatchSnapshot();
+});
